Add jwt auth on backend

+++ backend/src/main/java/nikolaichuks/teleconnect/backend/controller/TariffController.java
@@ -1,9 +1,13 @@
 package nikolaichuks.teleconnect.backend.controller;
 
 import lombok.RequiredArgsConstructor;
+import nikolaichuks.teleconnect.backend.model.User;
+import nikolaichuks.teleconnect.backend.model.UserRole;
 import nikolaichuks.teleconnect.backend.service.TariffService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.RestController;
 import teleconnect.tariff.api.TariffApi;
 import teleconnect.tariff.model.TariffDTO;
@@ -43,33 +47,33 @@ public class TariffController implements TariffApi {
 
     @Override
     public ResponseEntity<TariffDTO> createTariff(TariffDTO tariffDTO) {
+        boolean isAuthenticated = isEmployee();
+        if (!isAuthenticated) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return new ResponseEntity<>(tariffService.createTariff(tariffDTO), HttpStatus.CREATED);
     }
 
     @Override
     public ResponseEntity<TariffDTO> updateTariff(TariffDTO tariffDTO) {
+        boolean isAuthenticated = isEmployee();
+        if (!isAuthenticated) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return ResponseEntity.ok(tariffService.updateTariff(tariffDTO));
     }
 
     @Override
     public ResponseEntity<Void> deleteTariff(Integer id) {
+        boolean isAuthenticated = isEmployee();
+        if (!isAuthenticated) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         tariffService.deleteTariff(id);
         return ResponseEntity.ok().build();
     }
 
-//    @Override
-//    public ResponseEntity<List<TariffDTO>> searchTariffs(
-//            Double priceMin,
-//            Double priceMax,
-//            Integer dataLimitMin,
-//            Integer dataLimitMax,
-//            Integer callMinutesMin,
-//            Integer callMinutesMax,
-//            Integer smsLimitMin,
-//            Integer smsLimitMax,
-//            Boolean isActive,
-//            Boolean isUsed
-//    ) {
-//        return
-//    }
+    private boolean isEmployee() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        return currentUser.getRole().equals(UserRole.ROLE_EMPLOYEE);
+    }
+
 }
